Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/inline-event-HTML-entities-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/inline-event-HTML-entities.html&q=%3Cimg%20src=1%20onerror=%26%2397%26%23108%26%23101%26%23114%26%23116%26%2340%26%2349%26%2341%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
